requirements.txt
opencv-python
pillow
numpy
mss

pynput